pyproject.toml
[tool.poetry]
name = "hawa-be"
version = "0.1.0"
description = "FastAPI backend (auth + Postgres)"
authors = ["Arya <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "0.123.0"
uvicorn = {version = "0.38.0", extras = ["standard"]}
SQLAlchemy = "2.0.44"
psycopg = {version = "3.2.13", extras = ["binary"]}
python-jose = "3.5.0"
passlib = {version = "1.7.4", extras = ["bcrypt"]}
pydantic = {extras = ["email"], version = "^2.12.5"}
python-dotenv = "1.2.1"
httpx = "^0.27.2"
python-multipart = "^0.0.12"
apscheduler = "^3.10.4"

# Weather & LLM dependencies
groq = "^0.37.1"
pywhatkit = "^5.4"
pgvector = "^0.4.1"
pandas = "^2.3.3"
numpy = "2.3.5"
sentence-transformers = "^5.1.2"  # Untuk embeddings lokal
cryptography = "^46.0.3"  # Untuk encryption
gspread = "^6.2.1"  # Google Sheets API
# Keep google-auth in the range supported by google-auth-oauthlib (<2.42)
google-auth = "^2.41.0"  # Google Authentication
google-auth-oauthlib = "^1.2.3"
google-auth-httplib2 = "^0.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
pytest-asyncio = "0.24.0"
ruff = "0.6.9"
mypy = "1.13.0"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
